display profile sorted in new folder
