--- src/js/components/urlParams.js
@@ -1,9 +1,0 @@
-// /**
-//  * Utility function to get a specific query parameter from the URL.
-//  * @param {string} key - The key of the query parameter to retrieve.
-//  * @returns {string|null} - The value of the query parameter or null if not found.
-//  */
-// export function getQueryParam(key) {
-//   const urlParams = new URLSearchParams(window.location.search);
-//   return urlParams.get(key);
-// }
